+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Fixed
+- Parse errors where always reported to be in the header file that was input to the tool
+  rather than the actual, possibly included, file where the parse error occurred.
+
+
 ## [0.2.0] - 2025-08-26
 
 ### Added